RUB-248 rust: split spend verify shape
