+++ clients/rust/crates/rubin-consensus/src/spend_verify.rs
@@ -3,7 +3,9 @@ use crate::error::{ErrorCode, TxError};
 use crate::hash::sha3_256;
 use crate::sig_queue::{queue_or_verify_signature, SigCheckQueue};
 use crate::sighash::{is_valid_sighash_type, sighash_v1_digest_with_cache, SighashV1PrehashCache};
-use crate::suite_registry::{DefaultRotationProvider, RotationProvider, SuiteRegistry};
+use crate::suite_registry::{
+    DefaultRotationProvider, NativeSuiteSet, RotationProvider, SuiteParams, SuiteRegistry,
+};
 use crate::tx::Tx;
 use crate::tx::WitnessItem;
 use crate::utxo_basic::UtxoEntry;
@@ -118,39 +120,16 @@ pub(crate) fn validate_p2pk_spend_q(
     let default_reg = SuiteRegistry::default_registry();
     let rp: &dyn RotationProvider = rotation.unwrap_or(&default_rp);
     let reg = registry.unwrap_or(&default_reg);
-
     let native_spend = rp.native_spend_suites(block_height);
-    if !native_spend.contains(w.suite_id) {
-        return Err(TxError::new(
-            ErrorCode::TxErrSigAlgInvalid,
-            "CORE_P2PK suite not in native spend set",
-        ));
-    }
-
-    let params = reg.lookup(w.suite_id).ok_or_else(|| {
-        TxError::new(
-            ErrorCode::TxErrSigAlgInvalid,
-            "CORE_P2PK suite not registered",
-        )
-    })?;
-
-    if w.pubkey.len() as u64 != params.pubkey_len || w.signature.len() as u64 != params.sig_len + 1
-    {
-        return Err(TxError::new(
-            ErrorCode::TxErrSigNoncanonical,
-            "non-canonical witness item lengths",
-        ));
-    }
-    if entry.covenant_data.len() as u64 != MAX_P2PK_COVENANT_DATA
-        || entry.covenant_data[0] != w.suite_id
-    {
-        return Err(TxError::new(
-            ErrorCode::TxErrCovenantTypeInvalid,
-            "CORE_P2PK covenant_data invalid",
-        ));
-    }
-    let mut key_id = [0u8; 32];
-    key_id.copy_from_slice(&entry.covenant_data[1..33]);
+    let params = spend_suite_params(
+        w,
+        &native_spend,
+        reg,
+        "CORE_P2PK suite not in native spend set",
+        "CORE_P2PK suite not registered",
+    )?;
+    validate_witness_item_lengths(w, params)?;
+    let key_id = p2pk_covenant_key_id(entry, w.suite_id)?;
     let mut sig_queue = sig_queue;
     verify_mldsa_key_and_sig_q(
         w,
@@ -164,6 +143,49 @@ pub(crate) fn validate_p2pk_spend_q(
         TxError::new(ErrorCode::TxErrSigInvalid, "CORE_P2PK key binding mismatch"),
         TxError::new(ErrorCode::TxErrSigInvalid, "CORE_P2PK signature invalid"),
     )
+}
+
+fn spend_suite_params<'a>(
+    w: &WitnessItem,
+    native_spend: &NativeSuiteSet,
+    registry: &'a SuiteRegistry,
+    suite_not_native_message: &'static str,
+    suite_not_registered_message: &'static str,
+) -> Result<&'a SuiteParams, TxError> {
+    if !native_spend.contains(w.suite_id) {
+        return Err(TxError::new(
+            ErrorCode::TxErrSigAlgInvalid,
+            suite_not_native_message,
+        ));
+    }
+    registry
+        .lookup(w.suite_id)
+        .ok_or_else(|| TxError::new(ErrorCode::TxErrSigAlgInvalid, suite_not_registered_message))
+}
+
+fn validate_witness_item_lengths(w: &WitnessItem, params: &SuiteParams) -> Result<(), TxError> {
+    if w.pubkey.len() as u64 == params.pubkey_len && w.signature.len() as u64 == params.sig_len + 1
+    {
+        return Ok(());
+    }
+    Err(TxError::new(
+        ErrorCode::TxErrSigNoncanonical,
+        "non-canonical witness item lengths",
+    ))
+}
+
+fn p2pk_covenant_key_id(entry: &UtxoEntry, suite_id: u8) -> Result<[u8; 32], TxError> {
+    if entry.covenant_data.len() as u64 != MAX_P2PK_COVENANT_DATA
+        || entry.covenant_data[0] != suite_id
+    {
+        return Err(TxError::new(
+            ErrorCode::TxErrCovenantTypeInvalid,
+            "CORE_P2PK covenant_data invalid",
+        ));
+    }
+    let mut key_id = [0u8; 32];
+    key_id.copy_from_slice(&entry.covenant_data[1..33]);
+    Ok(key_id)
 }
 
 #[allow(dead_code)]
@@ -346,34 +368,17 @@ pub(crate) fn validate_threshold_sig_spend_q(
                 continue;
             }
 
-            if !native_spend.contains(w.suite_id) {
-                return Err(TxError::new(
-                    ErrorCode::TxErrSigAlgInvalid,
-                    threshold_sig_error_message(
-                        context,
-                        ThresholdSigErrorDetail::SuiteNotInNativeSpendSet,
-                    ),
-                ));
-            }
-
-            let params = reg.lookup(w.suite_id).ok_or_else(|| {
-                TxError::new(
-                    ErrorCode::TxErrSigAlgInvalid,
-                    threshold_sig_error_message(
-                        context,
-                        ThresholdSigErrorDetail::SuiteNotRegistered,
-                    ),
-                )
-            })?;
-
-            if w.pubkey.len() as u64 != params.pubkey_len
-                || w.signature.len() as u64 != params.sig_len + 1
-            {
-                return Err(TxError::new(
-                    ErrorCode::TxErrSigNoncanonical,
-                    "non-canonical witness item lengths",
-                ));
-            }
+            let params = spend_suite_params(
+                w,
+                &native_spend,
+                reg,
+                threshold_sig_error_message(
+                    context,
+                    ThresholdSigErrorDetail::SuiteNotInNativeSpendSet,
+                ),
+                threshold_sig_error_message(context, ThresholdSigErrorDetail::SuiteNotRegistered),
+            )?;
+            validate_witness_item_lengths(w, params)?;
 
             verify_mldsa_key_and_sig_q(
                 w,
@@ -449,1217 +454,5 @@ pub(crate) fn verify_mldsa_key_and_sig_q(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::constants::{
-        COV_TYPE_P2PK, ML_DSA_87_PUBKEY_BYTES, ML_DSA_87_SIG_BYTES, SIGHASH_ALL,
-        SUITE_ID_ML_DSA_87, SUITE_ID_SENTINEL, VERIFY_COST_ML_DSA_87,
-    };
-    use crate::hash::sha3_256;
-    use crate::sighash_v1_digest_with_cache;
-    use crate::suite_registry::{NativeSuiteSet, RotationProvider, SuiteParams, SuiteRegistry};
-    use crate::tx::{Tx, TxInput, TxOutput, WitnessItem};
-    use crate::verify_sig_openssl::Mldsa87Keypair;
-    use crate::SighashV1PrehashCache;
-    use std::collections::BTreeMap;
-
-    struct TestSpendSetRotation {
-        spend: NativeSuiteSet,
-    }
-
-    impl RotationProvider for TestSpendSetRotation {
-        fn native_create_suites(&self, _height: u64) -> NativeSuiteSet {
-            self.spend.clone()
-        }
-
-        fn native_spend_suites(&self, _height: u64) -> NativeSuiteSet {
-            self.spend.clone()
-        }
-    }
-
-    fn dummy_entry() -> UtxoEntry {
-        UtxoEntry {
-            value: 1,
-            covenant_type: COV_TYPE_P2PK,
-            covenant_data: vec![0u8; 100],
-            creation_height: 0,
-            created_by_coinbase: false,
-        }
-    }
-
-    fn dummy_tx_ctx() -> (Tx, u32, u64, [u8; 32]) {
-        let mut prev = [0u8; 32];
-        prev[0] = 0x42;
-        let mut chain_id = [0u8; 32];
-        chain_id[0] = 0x11;
-        (
-            Tx {
-                version: 1,
-                tx_kind: 0x00,
-                tx_nonce: 7,
-                inputs: vec![TxInput {
-                    prev_txid: prev,
-                    prev_vout: 0,
-                    script_sig: vec![],
-                    sequence: 0,
-                }],
-                outputs: vec![TxOutput {
-                    value: 1,
-                    covenant_type: COV_TYPE_P2PK,
-                    covenant_data: vec![0u8; 33],
-                }],
-                locktime: 0,
-                witness: vec![],
-                da_payload: vec![],
-                da_commit_core: None,
-                da_chunk_core: None,
-            },
-            0,
-            1,
-            chain_id,
-        )
-    }
-
-    fn sign_witness(
-        keypair: &Mldsa87Keypair,
-        tx: &Tx,
-        input_index: u32,
-        input_value: u64,
-        chain_id: [u8; 32],
-    ) -> WitnessItem {
-        let mut cache = SighashV1PrehashCache::new(tx).expect("cache");
-        let digest = sighash_v1_digest_with_cache(
-            &mut cache,
-            input_index,
-            input_value,
-            chain_id,
-            SIGHASH_ALL,
-        )
-        .expect("digest");
-        let mut signature = keypair.sign_digest32(digest).expect("sign");
-        signature.push(SIGHASH_ALL);
-        WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: keypair.pubkey_bytes(),
-            signature,
-        }
-    }
-
-    fn make_p2pk_entry(pubkey: &[u8]) -> UtxoEntry {
-        let pubkey_hash = sha3_256(pubkey);
-        let mut covenant_data = vec![SUITE_ID_ML_DSA_87];
-        covenant_data.extend_from_slice(&pubkey_hash);
-        UtxoEntry {
-            value: 100,
-            covenant_type: COV_TYPE_P2PK,
-            covenant_data,
-            creation_height: 0,
-            created_by_coinbase: false,
-        }
-    }
-
-    #[test]
-    fn extract_crypto_sig_and_sighash_rejects_missing_trailer() {
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![],
-            signature: vec![],
-        };
-        let err = extract_crypto_sig_and_sighash(&w).expect_err("missing sighash trailer");
-        assert_eq!(err.code, ErrorCode::TxErrParse);
-    }
-
-    #[test]
-    fn extract_crypto_sig_and_sighash_rejects_invalid_type() {
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![],
-            signature: vec![0xAA, 0xFF],
-        };
-        let err = extract_crypto_sig_and_sighash(&w).expect_err("invalid sighash type");
-        assert_eq!(err.code, ErrorCode::TxErrSighashTypeInvalid);
-    }
-
-    #[test]
-    fn extract_crypto_sig_and_sighash_preserves_crypto_sig_bytes() {
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![],
-            signature: vec![0xAA, 0xBB, SIGHASH_ALL],
-        };
-        let (crypto_sig, sighash_type) =
-            extract_crypto_sig_and_sighash(&w).expect("valid trailing sighash");
-        assert_eq!(crypto_sig, &[0xAA, 0xBB]);
-        assert_eq!(sighash_type, SIGHASH_ALL);
-    }
-
-    // ======== Registry & Lookup Tests (8) ========
-
-    #[test]
-    fn verify_sig_with_registry_nil_uses_default_live_registry() {
-        // When registry is None, verify_sig_with_registry should use the
-        // canonical default live registry rather than a separate legacy path.
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let pubkey = keypair.pubkey_bytes();
-        let digest = [0x42; 32];
-        let sig = keypair.sign_digest32(digest).expect("sign");
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            SUITE_ID_ML_DSA_87,
-            &pubkey,
-            &sig,
-            &digest,
-            None,
-        )
-        .expect("verify");
-
-        assert!(
-            result,
-            "valid signature should verify via default live registry"
-        );
-    }
-
-    #[test]
-    fn verify_sig_with_registry_unknown_suite_returns_error() {
-        // Suite 0xFF not in registry
-        let registry = SuiteRegistry::default_registry();
-        let pubkey = vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize];
-        let sig = vec![0x02; ML_DSA_87_SIG_BYTES as usize];
-        let digest = [0x42; 32];
-
-        let err = crate::verify_sig_openssl::verify_sig_with_registry(
-            0xFF,
-            &pubkey,
-            &sig,
-            &digest,
-            Some(&registry),
-        )
-        .expect_err("unknown suite should error");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-    }
-
-    #[test]
-    fn verify_sig_with_registry_known_suite_wrong_lengths() {
-        // Suite registered but pubkey/sig lengths mismatch
-        let registry = SuiteRegistry::default_registry();
-        let pubkey = vec![0x01; 10]; // wrong length
-        let sig = vec![0x02; 10]; // wrong length
-        let digest = [0x42; 32];
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            SUITE_ID_ML_DSA_87,
-            &pubkey,
-            &sig,
-            &digest,
-            Some(&registry),
-        )
-        .expect("should not error on length check");
-
-        assert!(!result, "wrong lengths should return false, not error");
-    }
-
-    #[test]
-    fn verify_sig_with_registry_known_suite_correct_lengths() {
-        // Valid suite with correct lengths passes through to OpenSSL
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let pubkey = keypair.pubkey_bytes();
-        let digest = [0x42; 32];
-        let sig = keypair.sign_digest32(digest).expect("sign");
-        let registry = SuiteRegistry::default_registry();
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            SUITE_ID_ML_DSA_87,
-            &pubkey,
-            &sig,
-            &digest,
-            Some(&registry),
-        )
-        .expect("verify with registry");
-
-        assert!(result, "valid signature with registry should verify");
-    }
-
-    #[test]
-    fn verify_sig_with_registry_custom_suite() {
-        // Custom registry entry with different suite ID
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let pubkey = keypair.pubkey_bytes();
-        let digest = [0x42; 32];
-        let sig = keypair.sign_digest32(digest).expect("sign");
-
-        let mut suites = BTreeMap::new();
-        suites.insert(
-            0x05,
-            SuiteParams {
-                suite_id: 0x05,
-                pubkey_len: ML_DSA_87_PUBKEY_BYTES,
-                sig_len: ML_DSA_87_SIG_BYTES,
-                verify_cost: VERIFY_COST_ML_DSA_87,
-                alg_name: "ML-DSA-87",
-            },
-        );
-        let registry = SuiteRegistry::with_suites(suites);
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            0x05,
-            &pubkey,
-            &sig,
-            &digest,
-            Some(&registry),
-        )
-        .expect("verify custom suite");
-
-        assert!(result, "custom suite entry should verify");
-    }
-
-    #[test]
-    fn verify_sig_with_registry_rejects_unsupported_binding() {
-        let mut suites = BTreeMap::new();
-        suites.insert(
-            0x05,
-            SuiteParams {
-                suite_id: 0x05,
-                pubkey_len: 1312,
-                sig_len: 2420,
-                verify_cost: 4,
-                alg_name: "ML-DSA-65",
-            },
-        );
-        let registry = SuiteRegistry::with_suites(suites);
-        let pubkey = vec![0x01; 1312];
-        let sig = vec![0x02; 2420];
-        let digest = [0x42; 32];
-
-        let err = crate::verify_sig_openssl::verify_sig_with_registry(
-            0x05,
-            &pubkey,
-            &sig,
-            &digest,
-            Some(&registry),
-        )
-        .expect_err("unsupported binding should fail closed");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-        assert!(err.msg.contains("unsupported suite verifier binding"));
-    }
-
-    #[test]
-    fn verify_sig_with_registry_consensus_init_error() {
-        // Empty inputs should trigger OpenSSL parse error
-        let registry = SuiteRegistry::default_registry();
-        let digest = [0x42; 32];
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            SUITE_ID_ML_DSA_87,
-            &[],
-            &[],
-            &digest,
-            Some(&registry),
-        );
-
-        // Empty inputs return false, not an error (length check happens first)
-        let ok = result.expect("should not error");
-        assert!(!ok, "empty inputs should fail verification");
-    }
-
-    #[test]
-    fn verify_sig_with_registry_openssl_error() {
-        // Corrupted signature should fail verification
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let pubkey = keypair.pubkey_bytes();
-        let digest = [0x42; 32];
-        let mut sig = keypair.sign_digest32(digest).expect("sign");
-        sig[0] ^= 0xFF; // corrupt first byte
-        let registry = SuiteRegistry::default_registry();
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            SUITE_ID_ML_DSA_87,
-            &pubkey,
-            &sig,
-            &digest,
-            Some(&registry),
-        )
-        .expect("verify");
-
-        assert!(!result, "corrupted signature should not verify");
-    }
-
-    #[test]
-    fn verify_sig_with_registry_verify_returns_false() {
-        // Valid signature over different digest should fail
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let pubkey = keypair.pubkey_bytes();
-        let digest1 = [0x42; 32];
-        let sig = keypair.sign_digest32(digest1).expect("sign");
-        let mut digest2 = digest1;
-        digest2[0] ^= 0xFF;
-        let registry = SuiteRegistry::default_registry();
-
-        let result = crate::verify_sig_openssl::verify_sig_with_registry(
-            SUITE_ID_ML_DSA_87,
-            &pubkey,
-            &sig,
-            &digest2,
-            Some(&registry),
-        )
-        .expect("verify");
-
-        assert!(!result, "signature over different digest should not verify");
-    }
-
-    // ======== P2PK Spend at Height Tests (7) ========
-
-    #[test]
-    fn p2pk_at_height_nil_providers_falls_back() {
-        // No rotation, no registry -> use defaults (ML-DSA-87)
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let entry = make_p2pk_entry(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let result = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            None,
-        );
-
-        assert!(result.is_ok(), "default providers should verify valid P2PK");
-    }
-
-    #[test]
-    fn p2pk_at_height_suite_not_in_spend_set() {
-        // Suite 0xFF not in native spend set at height
-        let entry = dummy_entry();
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: 0xFF,
-            pubkey: vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize],
-            signature: vec![0x02; ML_DSA_87_SIG_BYTES as usize + 1],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("unknown suite");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-    }
-
-    #[test]
-    fn p2pk_at_height_wrong_lengths() {
-        // Correct suite, wrong pubkey/sig lengths
-        let entry = dummy_entry();
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![0x01; 10], // wrong length
-            signature: vec![0x02; 10],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("wrong lengths");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigNoncanonical);
-    }
-
-    #[test]
-    fn p2pk_at_height_valid_sig_success() {
-        // Full valid P2PK spend path
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let entry = make_p2pk_entry(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let result = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            None,
-        );
-
-        assert!(result.is_ok(), "valid P2PK should verify");
-    }
-
-    #[test]
-    fn p2pk_at_height_bad_covenant_data() {
-        // Malformed covenant_data
-        let entry = UtxoEntry {
-            value: 1,
-            covenant_type: COV_TYPE_P2PK,
-            covenant_data: vec![0x99; 10], // too short
-            creation_height: 0,
-            created_by_coinbase: false,
-        };
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize],
-            signature: vec![0x02; ML_DSA_87_SIG_BYTES as usize + 1],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("bad covenant");
-
-        assert_eq!(err.code, ErrorCode::TxErrCovenantTypeInvalid);
-    }
-
-    #[test]
-    fn p2pk_at_height_suite_not_registered() {
-        // Suite in spend set but not in registry
-        let mut covenant_data = vec![0xAA];
-        covenant_data.extend_from_slice(&[0x11; 32]);
-        let entry = UtxoEntry {
-            value: 1,
-            covenant_type: COV_TYPE_P2PK,
-            covenant_data,
-            creation_height: 0,
-            created_by_coinbase: false,
-        };
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-
-        // Create custom registry with only SUITE_ID_ML_DSA_87
-        let mut suites = BTreeMap::new();
-        suites.insert(
-            SUITE_ID_ML_DSA_87,
-            SuiteParams {
-                suite_id: SUITE_ID_ML_DSA_87,
-                pubkey_len: ML_DSA_87_PUBKEY_BYTES,
-                sig_len: ML_DSA_87_SIG_BYTES,
-                verify_cost: 8,
-                alg_name: "ML-DSA-87",
-            },
-        );
-        let registry = SuiteRegistry::with_suites(suites);
-
-        let w = WitnessItem {
-            suite_id: 0xAA, // suite in covenant but not registered
-            pubkey: vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize],
-            signature: vec![0x02; ML_DSA_87_SIG_BYTES as usize + 1],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            Some(&registry),
-        )
-        .expect_err("unregistered suite");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-    }
-
-    #[test]
-    fn p2pk_at_height_key_binding_mismatch() {
-        // sha3(pubkey) != covenant key_id
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let mut covenant_data = vec![SUITE_ID_ML_DSA_87];
-        covenant_data.extend_from_slice(&[0xFF; 32]); // wrong key_id
-        let entry = UtxoEntry {
-            value: 1,
-            covenant_type: COV_TYPE_P2PK,
-            covenant_data,
-            creation_height: 0,
-            created_by_coinbase: false,
-        };
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_p2pk_spend_at_height(
-            &entry,
-            &w,
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("key binding mismatch");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-    }
-
-    // ======== Threshold Sig at Height Tests (10) ========
-
-    #[test]
-    fn threshold_at_height_nil_providers_falls_back() {
-        // No rotation, no registry -> defaults
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let result = validate_threshold_sig_spend_at_height(
-            &[key_id],
-            1,
-            &[w],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        );
-
-        assert!(result.is_ok(), "1-of-1 valid threshold should verify");
-    }
-
-    #[test]
-    fn threshold_at_height_sentinel_passthrough() {
-        // SENTINEL suite should be skipped (keyless)
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w1 = WitnessItem {
-            suite_id: SUITE_ID_SENTINEL,
-            pubkey: vec![],
-            signature: vec![],
-        };
-        let w2 = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let result = validate_threshold_sig_spend_at_height(
-            &[key_id, key_id],
-            1,
-            &[w1, w2],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        );
-
-        assert!(result.is_ok(), "sentinel should be skipped");
-    }
-
-    #[test]
-    fn threshold_at_height_non_native_suite_rejects() {
-        // Suite 0xFF not in native spend set
-        let key_id = [0x11; 32];
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: 0xFF,
-            pubkey: vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize],
-            signature: vec![0x02; ML_DSA_87_SIG_BYTES as usize + 1],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[key_id],
-            1,
-            &[w],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("bad suite");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-    }
-
-    #[test]
-    fn threshold_at_height_slot_count_mismatch() {
-        // Different number of witnesses vs keys
-        let key_id = [0x11; 32];
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[key_id, key_id], // 2 keys
-            1,
-            &[], // 0 witnesses
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("mismatch");
-
-        assert_eq!(err.code, ErrorCode::TxErrParse);
-    }
-
-    #[test]
-    fn threshold_at_height_valid_sigs_meets_threshold() {
-        // 1-of-2 threshold with one valid signature
-        let keypair1 = Mldsa87Keypair::generate().expect("keypair1");
-        let keypair2 = Mldsa87Keypair::generate().expect("keypair2");
-        let key_id1 = sha3_256(&keypair1.pubkey_bytes());
-        let key_id2 = sha3_256(&keypair2.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w1 = sign_witness(&keypair1, &tx, input_index, input_value, chain_id);
-        let w2 = WitnessItem {
-            suite_id: SUITE_ID_SENTINEL,
-            pubkey: vec![],
-            signature: vec![],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let result = validate_threshold_sig_spend_at_height(
-            &[key_id1, key_id2],
-            1,
-            &[w1, w2],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        );
-
-        assert!(result.is_ok(), "1-of-2 with one valid should pass");
-    }
-
-    #[test]
-    fn threshold_at_height_threshold_not_met() {
-        // Threshold 2 but only 1 valid signature
-        let keypair1 = Mldsa87Keypair::generate().expect("keypair1");
-        let keypair2 = Mldsa87Keypair::generate().expect("keypair2");
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let mut bad_sig = sign_witness(&keypair1, &tx, input_index, input_value, chain_id);
-        bad_sig.signature[0] ^= 0xFF; // corrupt signature
-        let w2 = sign_witness(&keypair2, &tx, input_index, input_value, chain_id);
-
-        let key_id1 = sha3_256(&keypair1.pubkey_bytes());
-        let key_id2 = sha3_256(&keypair2.pubkey_bytes());
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[key_id1, key_id2],
-            2,
-            &[bad_sig, w2],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("threshold not met");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-    }
-
-    #[test]
-    fn threshold_at_height_sentinel_with_payload_rejects() {
-        // SENTINEL suite with non-empty pubkey/signature
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: SUITE_ID_SENTINEL,
-            pubkey: vec![0x01], // sentinel must be keyless
-            signature: vec![],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0x11; 32]],
-            1,
-            &[w],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("sentinel payload");
-
-        assert_eq!(err.code, ErrorCode::TxErrParse);
-    }
-
-    #[test]
-    fn threshold_at_height_wrong_lengths() {
-        // Non-canonical witness item lengths
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![0x01; 10], // wrong length
-            signature: vec![0x02; 10],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0x11; 32]],
-            1,
-            &[w],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("wrong lengths");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigNoncanonical);
-    }
-
-    #[test]
-    fn threshold_at_height_not_registered() {
-        // Suite not in registry
-        let mut suites = BTreeMap::new();
-        suites.insert(
-            SUITE_ID_ML_DSA_87,
-            SuiteParams {
-                suite_id: SUITE_ID_ML_DSA_87,
-                pubkey_len: ML_DSA_87_PUBKEY_BYTES,
-                sig_len: ML_DSA_87_SIG_BYTES,
-                verify_cost: 8,
-                alg_name: "ML-DSA-87",
-            },
-        );
-        let registry = SuiteRegistry::with_suites(suites);
-
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: 0xBB, // not in registry
-            pubkey: vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize],
-            signature: vec![0x02; ML_DSA_87_SIG_BYTES as usize + 1],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0x11; 32]],
-            1,
-            &[w],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            Some(&registry),
-        )
-        .expect_err("unregistered suite");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-    }
-
-    #[test]
-    fn threshold_at_height_sig_verify_error() {
-        // Corrupted signature should fail verification
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let mut w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-        w.signature[0] ^= 0xFF; // corrupt
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let err = validate_threshold_sig_spend_at_height(
-            &[key_id],
-            1,
-            &[w],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "test",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("bad signature");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-    }
-
-    #[test]
-    fn threshold_at_height_core_context_messages_match_go_surface() {
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let suite_not_native = WitnessItem {
-            suite_id: 0xBB,
-            pubkey: vec![],
-            signature: vec![],
-        };
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0x11; 32]],
-            1,
-            &[suite_not_native],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "CORE_MULTISIG",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("suite not in native spend set");
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-        assert_eq!(err.msg, "CORE_MULTISIG suite not in native spend set");
-        let suite_not_registered = WitnessItem {
-            suite_id: 0xBB,
-            pubkey: vec![0x01; ML_DSA_87_PUBKEY_BYTES as usize],
-            signature: vec![0x02; ML_DSA_87_SIG_BYTES as usize + 1],
-        };
-        let only_unregistered = TestSpendSetRotation {
-            spend: NativeSuiteSet::new(&[0xBB]),
-        };
-        let default_registry = SuiteRegistry::default_registry();
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0x11; 32]],
-            1,
-            &[suite_not_registered],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "CORE_VAULT",
-            &mut cache,
-            Some(&only_unregistered),
-            Some(&default_registry),
-        )
-        .expect_err("suite not registered");
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-        assert_eq!(err.msg, "CORE_VAULT suite not registered");
-        let keypair_binding = Mldsa87Keypair::generate().expect("keypair binding");
-        let witness_binding =
-            sign_witness(&keypair_binding, &tx, input_index, input_value, chain_id);
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0xEE; 32]],
-            1,
-            &[witness_binding],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "CORE_MULTISIG",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("key binding mismatch");
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-        assert_eq!(err.msg, "CORE_MULTISIG key binding mismatch");
-        let kp_bad = Mldsa87Keypair::generate().expect("keypair invalid sig");
-        let mut witness_invalid_sig =
-            sign_witness(&kp_bad, &tx, input_index, input_value, chain_id);
-        witness_invalid_sig.signature[0] ^= 0xFF;
-        let key_id = sha3_256(&kp_bad.pubkey_bytes());
-        let err = validate_threshold_sig_spend_at_height(
-            &[key_id],
-            1,
-            &[witness_invalid_sig],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "CORE_VAULT",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("invalid signature");
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-        assert_eq!(err.msg, "CORE_VAULT signature invalid");
-        let threshold_not_met = WitnessItem {
-            suite_id: SUITE_ID_SENTINEL,
-            pubkey: vec![],
-            signature: vec![],
-        };
-        let err = validate_threshold_sig_spend_at_height(
-            &[[0x22; 32]],
-            1,
-            &[threshold_not_met],
-            &tx,
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "CORE_VAULT",
-            &mut cache,
-            None,
-            None,
-        )
-        .expect_err("threshold not met");
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-        assert_eq!(err.msg, "CORE_VAULT threshold not met");
-    }
-
-    #[test]
-    fn threshold_queue_flush_preserves_core_context_signature_invalid() {
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let mut witness = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-        witness.signature[0] ^= 0xFF;
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let registry = SuiteRegistry::default_registry();
-        let mut queue = crate::sig_queue::SigCheckQueue::new(1).with_registry(&registry);
-        validate_threshold_sig_spend_q(
-            &[key_id],
-            1,
-            &[witness],
-            input_index,
-            input_value,
-            chain_id,
-            0,
-            "CORE_MULTISIG",
-            &mut cache,
-            Some(&mut queue),
-            None,
-            Some(&registry),
-        )
-        .expect("queue path should defer signature verification");
-        let err = queue
-            .flush()
-            .expect_err("flush must surface invalid signature");
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-        assert_eq!(err.msg, "CORE_MULTISIG signature invalid");
-    }
-
-    // ======== Key & Sig with Registry Cache Tests (5) ========
-
-    #[test]
-    fn verify_key_sig_registry_cache_key_mismatch() {
-        // sha3(pubkey) != expected_key_id
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let wrong_key_id = [0xFF; 32];
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let registry = SuiteRegistry::default_registry();
-        let err = verify_mldsa_key_and_sig_q(
-            &w,
-            wrong_key_id,
-            input_index,
-            input_value,
-            chain_id,
-            &mut cache,
-            &registry,
-            &mut None,
-            TxError::new(ErrorCode::TxErrSigInvalid, "key mismatch"),
-            TxError::new(ErrorCode::TxErrSigInvalid, "sig invalid"),
-        )
-        .expect_err("key mismatch");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-    }
-
-    #[test]
-    fn verify_key_sig_registry_cache_sig_invalid() {
-        // Valid key but bad signature
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let mut w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-        w.signature[0] ^= 0xFF; // corrupt signature
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let registry = SuiteRegistry::default_registry();
-        let err = verify_mldsa_key_and_sig_q(
-            &w,
-            key_id,
-            input_index,
-            input_value,
-            chain_id,
-            &mut cache,
-            &registry,
-            &mut None,
-            TxError::new(ErrorCode::TxErrSigInvalid, "key mismatch"),
-            TxError::new(ErrorCode::TxErrSigInvalid, "sig invalid"),
-        )
-        .expect_err("bad sig");
-
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-    }
-
-    #[test]
-    fn verify_key_sig_registry_cache_openssl_error() {
-        // Empty pubkey should trigger key binding mismatch (sha3 of empty != key_id)
-        let key_id = [0x11; 32];
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: SUITE_ID_ML_DSA_87,
-            pubkey: vec![],
-            signature: vec![SIGHASH_ALL],
-        };
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let registry = SuiteRegistry::default_registry();
-        let err = verify_mldsa_key_and_sig_q(
-            &w,
-            key_id,
-            input_index,
-            input_value,
-            chain_id,
-            &mut cache,
-            &registry,
-            &mut None,
-            TxError::new(ErrorCode::TxErrSigInvalid, "key mismatch"),
-            TxError::new(ErrorCode::TxErrSigInvalid, "sig invalid"),
-        )
-        .expect_err("key binding mismatch on empty pubkey");
-
-        // Empty pubkey SHA3 won't match [0x11; 32], so we get key mismatch error
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-    }
-
-    #[test]
-    fn verify_key_sig_registry_cache_bad_sighash() {
-        // Invalid sighash_type byte
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let mut w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-        let last_idx = w.signature.len() - 1;
-        w.signature[last_idx] = 0xFF; // invalid sighash_type
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let registry = SuiteRegistry::default_registry();
-        let err = verify_mldsa_key_and_sig_q(
-            &w,
-            key_id,
-            input_index,
-            input_value,
-            chain_id,
-            &mut cache,
-            &registry,
-            &mut None,
-            TxError::new(ErrorCode::TxErrSigInvalid, "key mismatch"),
-            TxError::new(ErrorCode::TxErrSigInvalid, "sig invalid"),
-        )
-        .expect_err("bad sighash");
-
-        assert_eq!(err.code, ErrorCode::TxErrSighashTypeInvalid);
-    }
-
-    #[test]
-    fn verify_key_sig_registry_cache_success() {
-        // Full valid roundtrip: correct key, valid signature
-        let keypair = Mldsa87Keypair::generate().expect("keypair");
-        let key_id = sha3_256(&keypair.pubkey_bytes());
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = sign_witness(&keypair, &tx, input_index, input_value, chain_id);
-
-        let mut cache = SighashV1PrehashCache::new(&tx).expect("cache");
-        let registry = SuiteRegistry::default_registry();
-        let result = verify_mldsa_key_and_sig_q(
-            &w,
-            key_id,
-            input_index,
-            input_value,
-            chain_id,
-            &mut cache,
-            &registry,
-            &mut None,
-            TxError::new(ErrorCode::TxErrSigInvalid, "key mismatch"),
-            TxError::new(ErrorCode::TxErrSigInvalid, "sig invalid"),
-        );
-
-        assert!(result.is_ok(), "valid key and sig should verify");
-    }
-
-    #[test]
-    fn p2pk_suite_invalid_rejected_sig_alg_invalid() {
-        let entry = dummy_entry();
-        let (tx, input_index, input_value, chain_id) = dummy_tx_ctx();
-        let w = WitnessItem {
-            suite_id: 0x02, // non-native / unknown suite
-            pubkey: vec![0x01],
-            signature: vec![0x01],
-        };
-        let err = validate_p2pk_spend(&entry, &w, &tx, input_index, input_value, chain_id, 0)
-            .unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrSigAlgInvalid);
-    }
-}
+#[path = "spend_verify_tests/mod.rs"]
+mod spend_verify_tests;
